:recycle: split custom streams into separate modules
