+++ src/graphql/acsys/datastream/filterdupes.rs
@@ -1,0 +1,144 @@
+use super::{global, DataStream};
+use futures::Stream;
+use futures_util::StreamExt;
+use std::{collections::HashMap, pin::Pin, task::Poll};
+
+// Forwards a stream of DataReply types, removing entries that have a
+// decreasing timestamp (i.e. data duplicated in archive and live data
+// streams.
+
+struct FilterDupes {
+    s: DataStream,
+    latest: HashMap<i32, f64>,
+}
+
+// Friendly function to wrap a stream with the FilterDupes stream.
+
+pub fn filter_dupes(s: DataStream) -> DataStream {
+    Box::pin(FilterDupes::new(s))
+}
+
+impl FilterDupes {
+    pub fn new(s: DataStream) -> Self {
+        FilterDupes {
+            s,
+            latest: HashMap::new(),
+        }
+    }
+}
+
+impl Stream for FilterDupes {
+    type Item = global::DataReply;
+
+    fn poll_next(
+        mut self: Pin<&mut Self>, ctxt: &mut std::task::Context<'_>,
+    ) -> Poll<Option<Self::Item>> {
+        loop {
+            match self.s.poll_next_unpin(ctxt) {
+                Poll::Ready(Some(mut v)) => {
+                    // If we get an empty data packet, drop it.
+
+                    if v.data.is_empty() {
+                        continue;
+                    }
+
+                    let entry = self.latest.entry(v.ref_id).or_insert(0.0);
+
+                    // Find the starting point in the data in which the
+                    // timestamp is greater than the last one seen.
+
+                    let start_index = v.data[..]
+                        .partition_point(|info| info.timestamp <= *entry);
+
+                    // Update the last seen timestamp. WE NEED TO DO THIS
+                    // BEFORE DRAINING ANY DUPLICATES. You might think we
+                    // can always use `v.data.last()`, but it could be the
+                    // case that, after draining, there's no elements in the
+                    // vector so we do this while we know it's still safe.
+
+                    *entry = entry.max(v.data.last().unwrap().timestamp);
+
+                    // Remove any readings that have already been sent.
+
+                    v.data.drain(..start_index);
+                    if v.data.is_empty() {
+                        continue;
+                    }
+
+                    break Poll::Ready(Some(v));
+                }
+                v @ Poll::Ready(None) => break v,
+                v @ Poll::Pending => break v,
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::global;
+
+    fn data_info(ts: f64) -> global::DataInfo {
+        global::DataInfo {
+            timestamp: ts,
+            result: global::DataType::Scalar(global::Scalar {
+                scalar_value: ts / 2.0,
+            }),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_dedupe() {
+        use futures::stream::{self, StreamExt};
+
+        let input = &[
+            // device channel 0 receives two data points. These should
+            // go through.
+            global::DataReply {
+                ref_id: 0,
+                data: vec![data_info(100.0), data_info(110.0)],
+            },
+            // Another data point for device 0. This has the same timestamp
+            // as the previous so it shouldn't appear in the output.
+            global::DataReply {
+                ref_id: 0,
+                data: vec![data_info(110.0)],
+            },
+            // A different device has a data point. It should go through.
+            global::DataReply {
+                ref_id: 1,
+                data: vec![data_info(100.0)],
+            },
+            // Shouldn't return the first element.
+            global::DataReply {
+                ref_id: 0,
+                data: vec![data_info(105.0), data_info(115.0)],
+            },
+        ];
+        let mut s = super::filter_dupes(
+            Box::pin(stream::iter(input.clone())) as super::DataStream
+        );
+
+        assert_eq!(
+            s.next().await.unwrap(),
+            global::DataReply {
+                ref_id: 0,
+                data: vec![data_info(100.0), data_info(110.0),]
+            },
+        );
+        assert_eq!(
+            s.next().await.unwrap(),
+            global::DataReply {
+                ref_id: 1,
+                data: vec![data_info(100.0),]
+            },
+        );
+        assert_eq!(
+            s.next().await.unwrap(),
+            global::DataReply {
+                ref_id: 0,
+                data: vec![data_info(115.0),]
+            },
+        );
+    }
+}
